Fix bugs, more refactorings
