+++ pages/api/log.tsx
@@ -1,4 +1,3 @@
-import { Request, Response } from 'express';
 import * as fs from 'fs';
 import getConfig from 'next/config';
 
@@ -6,7 +5,7 @@ const { serverRuntimeConfig } = getConfig();
 
 const LOG_PATH = serverRuntimeConfig.REACT_APP_LOG_FILEPATH;
 
-export default async function handler(req: Request, res: Response) {
+export default async function handler(req: any, res: any) {
   if (!LOG_PATH) {
     res.status(200).send('Log filepath not specified.');
     return;
