+++ src/components/Footer/Footer.module.css
@@ -6,6 +6,7 @@
 .footerinfotext {
   margin: 25px auto;
   max-width: 860px;
+  color:#555;
 }
 
 .footer {
